Rename types.ts to common.ts, since it is no longer just types
