+++ src/index.ts
@@ -6,4 +6,4 @@ export { faceToSvgString } from "./faceToSvgString.js";
 export { default as svgs } from "./svgs.js";
 export { svgsIndex } from "./svgs-index.js";
 
-export type * from "./types.js";
+export type * from "./common.js";
